Complete rewrite
Turns out that ovebrounces only depend on the cumulative frametimes.
So there is no need to run many simulated falls for a fixed fall
distance on recorded frametime history and remember success rates as
done previously.
Instead, we can create a statistic of the cumulative frametimes and use
those to immediately obtain the probability of any fall distance.